bugFix: empty primary application directory
This situation occurs when the repository's credentials are invalid. In this case the directory is created but no files are added.

The consequence is that the main directory is found, but the init file is not, which results in an exception when importing the file.
Finally, the unhandled exception results in the microcontroller reset.

Now, only creates the directory of the primary application if it has been successfully downloaded.